+++ CHANGELOG.md
@@ -1,10 +1,18 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## Current version
+## [3.0.2]
+
+### Features
+- ``UdsConfigTool``: handle parsing of dynamic response lengths in ODX files
+- ``CanTp``: Make reception polling interval configurable through the ``polling_interval`` attribute 
+
+### Changes
+- ``CanTp``: log a warning instead of raising an error instead of any unexpected received message
 
 ### Bugfixes
-- Fix padding of consecutive frames over CAN
+- ``CanTp``: Fix padding of consecutive frames over CAN
+- ``CanTp``: Avoid deletion of received CAN messages when sending a message
 
 ## [3.0.1]